HSEARCH-3162 Classes should not access their own subclasses during initialization
See https://sonarcloud.io/organizations/hibernate/rules?open=squid%3AS2390&rule_key=squid%3AS2390
