+++ backend-elasticsearch/src/main/java/org/hibernate/search/v6poc/backend/elasticsearch/gson/impl/JsonArrayAccessorImpl.java
@@ -23,7 +23,7 @@ public class JsonArrayAccessorImpl extends TypingJsonAccessor<JsonArray>
 
 	@Override
 	protected JsonElementType<JsonArray> getExpectedElementType() {
-		return JsonElementType.ARRAY;
+		return JsonElementTypes.ARRAY;
 	}
 
 	@Override
