+++ backend-elasticsearch/src/main/java/org/hibernate/search/v6poc/backend/elasticsearch/gson/impl/JsonDoubleAccessorImpl.java
@@ -19,7 +19,7 @@ public class JsonDoubleAccessorImpl extends TypingJsonAccessor<Double> {
 
 	@Override
 	protected JsonElementType<Double> getExpectedElementType() {
-		return JsonElementType.DOUBLE;
+		return JsonElementTypes.DOUBLE;
 	}
 
 }
