+++ backend-elasticsearch/src/main/java/org/hibernate/search/v6poc/backend/elasticsearch/gson/impl/JsonFloatAccessorImpl.java
@@ -19,7 +19,7 @@ public class JsonFloatAccessorImpl extends TypingJsonAccessor<Float> {
 
 	@Override
 	protected JsonElementType<Float> getExpectedElementType() {
-		return JsonElementType.FLOAT;
+		return JsonElementTypes.FLOAT;
 	}
 
 }
