+++ backend-elasticsearch/src/main/java/org/hibernate/search/v6poc/backend/elasticsearch/gson/impl/JsonIntegerAccessorImpl.java
@@ -19,7 +19,7 @@ public class JsonIntegerAccessorImpl extends TypingJsonAccessor<Integer> {
 
 	@Override
 	protected JsonElementType<Integer> getExpectedElementType() {
-		return JsonElementType.INTEGER;
+		return JsonElementTypes.INTEGER;
 	}
 
 }
