+++ backend-elasticsearch/src/main/java/org/hibernate/search/v6poc/backend/elasticsearch/gson/impl/JsonLongAccessorImpl.java
@@ -19,7 +19,7 @@ public class JsonLongAccessorImpl extends TypingJsonAccessor<Long> {
 
 	@Override
 	protected JsonElementType<Long> getExpectedElementType() {
-		return JsonElementType.LONG;
+		return JsonElementTypes.LONG;
 	}
 
 }
