+++ backend-elasticsearch/src/main/java/org/hibernate/search/v6poc/backend/elasticsearch/gson/impl/JsonObjectAccessorImpl.java
@@ -24,7 +24,7 @@ public class JsonObjectAccessorImpl extends TypingJsonAccessor<JsonObject>
 
 	@Override
 	protected JsonElementType<JsonObject> getExpectedElementType() {
-		return JsonElementType.OBJECT;
+		return JsonElementTypes.OBJECT;
 	}
 
 	@Override
