+++ backend-elasticsearch/src/main/java/org/hibernate/search/v6poc/backend/elasticsearch/types/codec/impl/IntegerFieldCodec.java
@@ -6,7 +6,7 @@
  */
 package org.hibernate.search.v6poc.backend.elasticsearch.types.codec.impl;
 
-import org.hibernate.search.v6poc.backend.elasticsearch.gson.impl.JsonElementType;
+import org.hibernate.search.v6poc.backend.elasticsearch.gson.impl.JsonElementTypes;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonNull;
@@ -33,6 +33,6 @@ public class IntegerFieldCodec implements ElasticsearchFieldCodec {
 		if ( element == null || element.isJsonNull() ) {
 			return null;
 		}
-		return JsonElementType.INTEGER.fromElement( element );
+		return JsonElementTypes.INTEGER.fromElement( element );
 	}
 }
